+++ test/transceivercontroller/SimulcastTransceiverController.test.ts
@@ -419,17 +419,18 @@ describe('SimulcastTransceiverController', () => {
 
       videosToReceive = new DefaultVideoStreamIdSet([7]);
       videoSubscriptions = tc.updateVideoTransceivers(videoStreamIndex, videosToReceive);
-      expect(videoSubscriptions).to.deep.equal([0, 7, 0]);
-      verifyTransceiverDirections(['recvonly', 'inactive']);
+      expect(videoSubscriptions).to.deep.equal([0, 0, 0, 7]);
+      verifyTransceiverDirections(['inactive', 'inactive', 'recvonly']);
       subackFrame = new SdkSubscribeAckFrame({
         tracks: [new SdkTrackMapping({ streamId: 7, trackLabel: 'v_7' })],
       });
       videoStreamIndex.integrateSubscribeAckFrame(subackFrame);
       setTransceiverStreamId(videoStreamIndex, videosToReceive);
+
       videosToReceive = new DefaultVideoStreamIdSet([7, 8]);
       videoSubscriptions = tc.updateVideoTransceivers(videoStreamIndex, videosToReceive);
-      expect(videoSubscriptions).to.deep.equal([0, 7, 8]);
-      verifyTransceiverDirections(['recvonly', 'recvonly']);
+      expect(videoSubscriptions).to.deep.equal([0, 0, 0, 7, 8]);
+      verifyTransceiverDirections(['inactive', 'inactive', 'recvonly', 'recvonly']);
     });
 
     it('will reuse transceiver for same group', () => {