+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- Fixed Firefox video tiles containing stale frames from previous transceivers by not attempting to reuse inactive transceivers.
 
 ### Changed